array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,7 @@
 #include "function_pointers.h"
+#include <stddef.h>
 /**
-  * array_iterator - executes a function given as a parameter 
+  * array_iterator - executes a function given as a parameter
   * on each element of array
   * @array: pointer to array
   * @size: size of array
@@ -10,7 +11,7 @@
   */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (!array || !size || !action)
 		return;
